Fundamental Kotlin - Inheritance.
